refactor: rename and move the files in parent dir
