+++ GANDLF/config_manager.py
@@ -3,9 +3,9 @@ from typing import Union
 import yaml
 from pydantic import ValidationError
 
-from GANDLF.Configuration.parameters_config import Parameters
-from GANDLF.Configuration.exclude_parameters import exclude_parameters
-from GANDLF.Configuration.utils import handle_configuration_errors
+from GANDLF.configuration.parameters_config import Parameters
+from GANDLF.configuration.exclude_parameters import exclude_parameters
+from GANDLF.configuration.utils import handle_configuration_errors
 
 
 def _parseConfig(
